+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release Notes
 All notable changes to this package will be documented in this file.
 
+# 0.5.1
 - Changed language files
 - Added an icon button menu in **AppBar**
 - Created `SwipeableDrawer` organism component in the **AppBar**